Fixes #2315 - Check if a config.xml file exist, before assuming it does.

+++ src/MythTV_Player/mkr_postinst.sh
@@ -7,11 +7,15 @@
 ln -s /usr/lib/libXmu.so.6.2.0 /usr/lib/libXmu.so || :
 # END  : Hack to get mythtv working
 
-# START: Hack to transpose xml
-echo $(grep -e '<Host>' /etc/mythtv/config.xml | sed -e 's#<Host>#DBHostName=#' -e 's#</Host>#\n#') > /etc/mythtv/mysql.txt
-echo $(grep -e '<UserName>' /etc/mythtv/config.xml | sed -e 's#<UserName>#DBUserName=#' -e 's#</UserName>#\n#') >> /etc/mythtv/mysql.txt
-echo $(grep -e '<DatabaseName>' /etc/mythtv/config.xml | sed -e 's#<DatabaseName>#DBName=#' -e 's#</DatabaseName>#\n#') >> /etc/mythtv/mysql.txt
-echo $(grep -e '<Password>' /etc/mythtv/config.xml | sed -e 's#<Password>#DBPassword=#' -e 's#</Password>#\n#') >> /etc/mythtv/mysql.txt
+# In case we have a later MythTV, we need to transpose the information
+# from the config.xml file.
+if [ -f /etc/mythtv/config.xml ] ; then
+	# START: Hack to transpose xml
+	echo $(grep -e '<Host>' /etc/mythtv/config.xml | sed -e 's#<Host>#DBHostName=#' -e 's#</Host>#\n#') > /etc/mythtv/mysql.txt
+	echo $(grep -e '<UserName>' /etc/mythtv/config.xml | sed -e 's#<UserName>#DBUserName=#' -e 's#</UserName>#\n#') >> /etc/mythtv/mysql.txt
+	echo $(grep -e '<DatabaseName>' /etc/mythtv/config.xml | sed -e 's#<DatabaseName>#DBName=#' -e 's#</DatabaseName>#\n#') >> /etc/mythtv/mysql.txt
+	echo $(grep -e '<Password>' /etc/mythtv/config.xml | sed -e 's#<Password>#DBPassword=#' -e 's#</Password>#\n#') >> /etc/mythtv/mysql.txt
+fi
 # END: Hack to transpose xml
 
 # If this is not the hybrid, copy the mysql.txt file from hybrid
